requirements.txt
openai
anthropic
python-dotenv
azure-cognitiveservices-speech